refactor: Separated likes into like_controller.py
